+++ onmt/modules/init.lua
@@ -3,8 +3,9 @@ onmt = onmt or {}
 require('onmt.modules.Sequencer')
 require('onmt.modules.Encoder')
 require('onmt.modules.BiEncoder')
-require('onmt.modules.DBiEncoder')
 require('onmt.modules.PDBiEncoder')
+require('onmt.modules.DBiEncoder')
+require('onmt.modules.GoogleEncoder')
 require('onmt.modules.Decoder')
 
 require('onmt.modules.Network')