+++ test/onmt/EncoderTest.lua
@@ -69,8 +69,6 @@ local function genericCheckMasking(encoder)
 
   local batch = onmt.data.Batch.new(src)
 
-  encoder:maskPadding()
-
   local _, context = encoder:forward(batch)
 
   tester:eq(context[2][1]:ne(0):sum(), 0)
@@ -183,6 +181,36 @@ end
 
 function encoderTest.dbrnn_saveAndLoad_GRU()
   local encoder, opt = buildEncoder(onmt.DBiEncoder, 'GRU')
+  genericCheckSerial(encoder, opt)
+end
+
+function encoderTest.gnmt_LSTM()
+  local encoder, opt = buildEncoder(onmt.GoogleEncoder, 'LSTM')
+  genericCheckDim(encoder, opt)
+end
+
+function encoderTest.gnmt_masking_LSTM()
+  local encoder, _ = buildEncoder(onmt.GoogleEncoder, 'LSTM')
+  genericCheckMasking(encoder)
+end
+
+function encoderTest.gnmt_saveAndLoad_LSTM()
+  local encoder, opt = buildEncoder(onmt.GoogleEncoder, 'LSTM')
+  genericCheckSerial(encoder, opt)
+end
+
+function encoderTest.gnmt_GRU()
+  local encoder, opt = buildEncoder(onmt.GoogleEncoder, 'GRU')
+  genericCheckDim(encoder, opt)
+end
+
+function encoderTest.gnmt_masking_GRU()
+  local encoder, _ = buildEncoder(onmt.GoogleEncoder, 'GRU')
+  genericCheckMasking(encoder)
+end
+
+function encoderTest.gnmt_saveAndLoad_GRU()
+  local encoder, opt = buildEncoder(onmt.GoogleEncoder, 'GRU')
   genericCheckSerial(encoder, opt)
 end