+++ train.lua
@@ -47,7 +47,7 @@ local function loadDataset(filename)
 
   -- Check if data type is compatible with the target model.
   if not modelClass.dataType(dataset.dataType) then
-    _G.logger:error('Data type `%s\' is incompatible model `%s\'',
+    _G.logger:error('Data type `%s\' is incompatible with `%s\' models',
                     dataset.dataType, modelClass.modelName())
     os.exit(0)
   end